+++ pygasflow/utils/doc_utils.py
@@ -203,7 +203,24 @@ def _modify_example_compressible_app(code):
     return new_code
 
 
+def _modify_diagram_show_only_figure(code):
+    lines = code.split("\n")
+    lines[-1] = "pn.pane.Bokeh(d.figure)"
+    new_code = "\n".join(lines)
+    return new_code
+
+def _modify_diagram_show_interactive_app(code):
+    lines = code.split("\n")
+    lines[-1] = "d = " + lines[-1]
+    lines.append("d.__panel__()")
+    new_code = "\n".join(lines)
+    return new_code
+
 def modify_panel_code(code):
     if "compressible_app" in code:
         return _modify_example_compressible_app(code)
+    elif "Diagram" in code:
+        if "bokeh.plotting" in code:
+            return _modify_diagram_show_only_figure(code)
+        return _modify_diagram_show_interactive_app(code)
     return code